LPS-138258 Adding a method to  InfoItemFieldValuesProvider interface.
The InfoFieldValues of templated InfoFields couldn't be included inside InfoItemFieldValues. That's because, to render an object template view, all object's InfoItemFieldValues must be included in the context, so if the template view is part of the needed InfoFieldValues then that will cause an infinite loop.

So, we need to put the InfoFieldValues of templated InfoFields off from InfoItemFieldValues and provide a method that allows us to get the value in a specific object of a templated info field.