LPS-143090 Context lazy initialization.
Sometimes when portal startup some info framework implementations are not completely registered when we try to register its menu item type, so the type is not registered because we can't reach the components we need to initialize the context.

In that way, we register one menu item type for each info framework implementation which has display page capability although the implementation is not fulfilled yet.