LPS-147251 Set explicit "none" as paginationType when upgrading if pagination is not set.
We need to retrieve the paginationType value directly from JSON data because the old paginationType default value was saved as null corresponding to "none" pagination.
As currently default value is "numeric", when creating a CollectionStyledLayoutStructureItem if no specific paginationType is set in the config JSON it takes its default value.
Then when upgrading turns to "numeric" the paginationType for those items which has null values.