LPS-150919 Adds new methods to get LayoutSEOLinks but manages the construction of alternates map instead of receiving it as a parameter.
It allows us to get the proper friendly URL for the target locale and use it in order to get the proper alternate. And to achieve it we are using the new class that does it in that way.

In the implementation, the logic changes just in the order to get the friendly URL before calculating the alternate URL. In that way, we are able to calculate the alternate URL using as canonical URL as the proper friendly URL for the locale.

It should be in that order because the proper alternate URL for a locale maybe is not the same that the proper friendly URL for that locale. That's because alternate URLs always include the i18n prefix except in the default locale but the friendly URL could include it or not depending on the configuration set in locale.prepend.friendly.url.style