LPS-165478 Preserve the option values instead of changing them for their labels. Otherwise, we are missing the option values in the Freemarker context, and it is useful, for example, in the information templates.
We are going to handle it at another point in order to be able to change the option value for its label only when we are mapping a select info field to a fragment. In that way, we do not affect the information templates that will have the option value as data and will be able to recover the label using the template node's options map.