LPS-188018 Adds a mock implementation for AICreatorOpenAIClient in order to provide a way to run manual and POSHI tests without interact with the real API.
We are providing a key that is considered as a valid API key and other keys that could be used to simulate the different handled errors.

The error keys provided are:

OPENAI_API_INVALID_API_KEY --> the response will be the same as when the real API returns invalid API key's response.
OPENAI_API_IOEXCEPTION --> the response will be the same as when the real API throws an exception.
OPENAI_API_{0}_RESPONSE_CODE --> the response will be the same as when the real API response code is {0}.