LPS-200228 Adds a ModelListener for LayoutPageTemplateEntry to prevent the content type changes when the display page is used by at least one asset.
I am using a model listener for this because trying to check it in the service layer will cause a circular dependency between AssetDisplayPageEntryLocalService and LayoutPageEntryLocalService. Checking it only in the action command would leave open the possibility to change the content type of a used display page by invoking directly the service method.