LPS-200225 Pass existsMappedContentType to DisplayPageActionDropdownItemsProvider and use it to prevent access directly to the edit action if the display page is mapped to a non-existing type or a subtype.
To do it, on edit action click, we are displaying the change content type popup, and redirect to the edition once the content-type is fixed.

So we are keeping the Edit action label but setting as dropdownItem action the changeContentType, In addition, at this point, we are setting as redirect in changeContentTypeURL the URL to edit the display page.